Posts and Comments with no text ignored
In response to ~ravmel's issue here: https://github.com/urbit/fora/issues/6

With this change if you try to submit a post or comment without text, then the poke is ignored.  Yes, this is a really dumb fix, but it works and fora is minimalist anyway.  Only downside is there is no explicit feedback to the poster.  If someone wants to hold my hand a little I'll try to make a more fancy change.