More accurately represent Ethereum RPC filter topics
Rather than a flat list of topics to watch for, it is an indexed list
of topics to watch for in those specific positions. As such, items in
that list may also be zero (null, matches any topic), or a list of
topics (matches any one of those topics).

Additionally, topics are always specified using 32 bytes, so there is
no need to specify length explicitly.