Replace +ob with a corrected version.
'ob', a core that implemented a permutation cipher for planet-sized
atoms, contained a critical bug that allowed multiple inputs to encipher
to the same output.  This created collisions between some planet-sized
@p values.

The rewritten version corrects this bug, and also rewrites +ob to more
clearly match both 1) the paper that the algorithm comes from, and 2)
the parallel implementation in urbit-ob that has gone through an
exhaustive test.

This also adjusts two calls to 'ob' elsewhere in hoon.hoon such that
they match the updated core.