Verify mugs on log replay to detect nondeterministic replay.
Joe helped write the worker/main.c changes so that we consistently
return the mug of the current state on each %done message. Then
enabled a check in _pier_work_complete() which asserts that the
recomputed event produces the same state as the next event was
expecting to be applied to.