Redo how we perform ECC math for ring signatures.
This makes three new functions in +ed:crypto which are designed to
be jetted by the ed25519 implementation. It then moves the ring
signature library stuff onto two of them. Previously, it took three
minutes to sign+verify an unlinked ring of 3 public keys; now it is
instantaneous with the jets written so far.

This comments out all the linkage handling code, since a jet isn't
written for that stuff yet.