Interface: Add "clean 'renamed' files" gulp task
gulp.rename just copies files, leaving the original.
For minified files, this was navigated by not
tracking the -min.js files in git. However, when
deploying OTA, the -min.js files are still deployed
to all ships. This commit adds 'del' package,
and a task to clean up the -min files when done
for all current Landscape applications.