build: use HEAD in sh/create-ropsten-pills
If we're not on a branch, e.g. if we've checked out an arbitrary
revision, then 'git branch --show-current' sensibly won't give us any
output, and thus the pills produced by sh/create-ropsten-pills will be
misnamed.

Consuming HEAD with git-rev-parse has the same effect here, and should
work in all cases of interst.