publish: add vertical padding to comment-field component
Prior to this commit, the commmenting component had a styling value of `mt9`, which spaced it properly away from the bottom of the post text, but rendered it with zero spacing from the bottom of the viewport in the case of no comments, and pushed right up against new comments or page navigation in an unsightly manner. This value has been edited to `mv9`, to give the comments component proper breathing room from other components.