metadata-store: never crash fetching app metadata
In order to track & respond to all app-related metadata, hooks may want to
subscribe to the metadata-store as the first thing they do on-init. In that case
it is exceedingly likely that there are no entries for the app-name in
app-indices yet.

This makes us fall back to the empty set in case app-indices has no entry for
the app-name yet.