Revert "Merge branch 'liam-fitzgerald/hoon-spot' (#2247)"
This reverts commit 2ab347ec333ddea4df63254e15cf8086208ec041, reversing
changes made to 5e8474de2f416440ffe475e7154d7262de52a806.

I'm reverting this as we're moving to a new branch/release model in
which breaching changes (as this one is) will live on a long-running
'next' branch, rather than alongside non-breaching changes in master.

This revert should itself be reverted on the 'next' branch.