chat: strictly match URLs in outgoing strings
Our regex processes true, and sends a message as a URL, if any part of
the string is a URL starting with a URI schema.
We pass this function strings cut up by spaces. If a link was enclosed
in quotes, or in brackets, this would still process as true.

This adds ^ to the regex to only process strings that START with
a schema.