chat-js: load all unreads and autoread on activity
if the number of unread messages is larger that the number we are going
to load, then load enough messages to display the unread marker.
Additionally, only automatically read a message if the user has been
active in the last minute. Freeze scroll position on inactivity.  Also
unconditionally scroll to the bottom upon sending your own message.