groups: fix race condition in joining flow
Attempting to sync a group that a ship is not in causes the subscription
to fail. Because the %add-members action and the sync used to be sent in
one invocation, ames does not guaranteed the order of these remote
actions and so we wait for the %add-members poke to be acked before
adding the group and associated resources.