vere: increase LMDB mapsize to 1TB on macOS, linux
Many ships have been observed bumping up against the existing mapsize
limits.  This results in a Vere crash via LMDB, which necessitates
compiling a new binary with a higher mapsize if one wants to relaunch.

There doesn't seem to be any serious penalty to setting this
somewhere in the terabyte range, though. [1]  In cases where the mapsize
exceeds the size of the disk, I infer from the LMDB docs that the
database may simply be permitted to grow until it runs up against the
disk limitations, which feels acceptable.

I've tested this on macOS and Linux and the binary runs without issue,
despite the mapsize being set far in excess of the disks I'm running on.

[1]: https://lmdb.readthedocs.io/en/release/