build: force google-cloud-sdk to use python3
This prevents an error where the python2 version of the SDK tries to
use the monotonic library which is unable to detect the host operating
system inside the sandbox - exploding in typical python fashion.

By using python3 it obviates the need for the monotonic dependency by
using the builtin time.monotonic call.