meta: pull in lfs files when using checkout action
Some other commits have been squashed in:

meta: remove user declarations for glob action

This is a tricky problem.  SSH needs UID 0 to exist, and intuitively it
does (root), but we can't enforce that when GitHub is running our
container.  Specifying "USER 0:0" in the actions Dockerfile doesn't cut
it.

Instead it seems best to try and construct a shadow-utils setup on the
underlying Docker image.

meta: target glob workflow at release/next-js