vere: remove trailing newlines from u3l_log calls
As of f0ceb560e these are redundant, and keeping them in would result
in trailing whiteline.

The result is that sometimes we call u3l_log with the empty string. We
simply want it to insert a newline into the scrollback for us. GCC
complains about this, because it leads to using an empty format string,
but we consider this sane behavior. We update the flags accordingly.