Inbox: make useLazyScroll account for archived notifications
useLazyScroll was not accounting for the case where a notification would
be archived. This would mean that archiving a lot of notifications
without scrolling would erroneously not load more. Accounts for this
case by passing a count argument to useLazyScroll, which is then
compared against its previous value. If the previous value is greater
than the new value, (i.e. a notification was archived) then check if we
should load more.