term: correctly (re)store styled lines
We were storing the contents of the bottom line, but not accounting for
style/escape characters when overwriting parts of it with new contents.

Here, we switch to storing a +stub noun in state instead, letting us
rely on the stdlib for the overwriting logic.
This comes at the cost of having to re-render the stub into an escape
sequence on every restore, but that has yet to prove itself slow enough
to care about.