docket: update docket file's glob hash on upload
If the docket says the glob is a local ames glob, then update the hash
to match the newly uploaded glob.

One would naively expect this to result in a %commit fact being
received, letting us update our state through the usual mechanism.
This is, for some reason, not the case...