build: do not include debug symbols in release
On M1 Macs, the compiler seems to infinite-loop, consuming ever more RAM
and CPU, while trying to build `noun/allocate.c` with `-O3 -g`. `-O3` is
fine, `-g` is fine. Both at once seems to try to summon demons.

Other possible solutions aside from this:

- try lower levels of optimization until we find one that doesn't hang,
  and ship that.

- do not support M1 Mac until the underlying issue here is fixed.

- ship debug binaries on M1 for now.

The path of least resistance is of course the second option, as that's
what is already tacitly happening.