clay: track pending (scry) requests better
Instead of having a separate "busy" flag and pending scry request state,
we now have a single "busy" unit that, if there is a pending request,
contains details about the pending request. In the ames case, this is
simply %ames. But in the scry case, it contains all the details we need
to cancel the request, timer, or both when needed.

Additionally, we now make sure to always cancel the scry timeout timer
whenever we get a scry response.