webterm: simpler printf rendering
As it turns out, the exact behavior of the 'S' CSI command is not "move
contents up into scrollback", but rather "delete the top line(s), move other
contents up". This behavior leads webterm to eat into outputs whenever it
renders slogs.

Xterm(.js), when given a newline character at the bottom of a scroll region,
does produce the desired behavior of bumping the top line away into scrollback,
instead of obliterating it.

This implementation now diverges from vere's, which is unfortunate, but the
alternative is making webterm track the contents of the bottom-most line of
the default session, which seems rather ridiculous by comparison.

Fixes #6257.